cleaned up target nics a bit
